back/pyproject.toml
[tool.poetry]
name = "django-fastapi-app"
version = "0.1.0"
description = ""
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.2"
django = "^5.1.2"
uvicorn = "^0.31.1"
django-environ = "^0.11.2"
tortoise-orm = "^0.21.7"

[tool.ruff]
ignore = ["F401"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
